First set of warning-silencing effort

+++ src/vector.cc
@@ -378,7 +378,7 @@ Vector Vector::fromGSLVector(const gsl_vector * vect)
 {
   Vector ret;
   ret.reserve(vect->size);
-  for(int i = 0; i < vect->size; i++)
+  for(size_t i = 0; i < vect->size; i++)
     ret << gsl_vector_get(vect, i);
   return ret;
 }
